online_shop/requirements.txt
asgiref==3.2.10
autopep8==1.5
bcrypt==3.1.7
cffi==1.14.0
cryptography==2.9
Django==3.1.2
django-extensions==3.0.9
django-filter==2.4.0
djangorestframework==3.12.1
Markdown==3.3.2
paramiko==2.0.2
Pillow==8.0.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.20
PyNaCl==1.3.0
pytz==2020.1
six==1.14.0
SQLAlchemy==1.3.16
sqlparse==0.4.1
